// src/Models/IFilePicker.ts
export interface IFilePickerOptions {
    sdk: '8.0';
    messaging?: IMessagingConfiguration;
    entry: IEntryConfiguration;
    telemetry?: ITelemetryConfiguration;
    authentication: IAuthenticationConfiguration;
    localization?: ILocalizationConfiguration;

    /**
     * Specifies what types of items may be picked and where they come from.
     */
    typesAndSources?: ITypesAndSourcesConfiguration;
    /**
     * Specified how many items may be picked.
     */
    selection?: ISelectionConfiguration;
    /**
     * Specifies what happens when users pick files and what the user may do with files in the picker.
     */
    commands?: ICommandConfiguration;
    /**
     * Specifies accessibility cues such as auto-focus behaviors.
     */
    accessibility?: IAccessibilityConfiguration;

    // TBD.
    navigation?: INavigationConfiguration;
    theme?: 'default' | 'dark' | 'lists' | 'teams' | 'teams-default' | 'teams-dark' | 'teams-contrast' | ITheme;
        
}

export interface IAuthenticateCommand {
    command: 'authenticate';
    resource: string;
    type: 'SharePoint' | 'Graph';
    claimsChallenge?: {
        claims: string;
        token?: string;
        error: any; // IError;
    };
    expiration?: {
        error: any; // IError;
    };
}
// ...
export interface INotificationData {
    code: string;
    isExpected: boolean;
    message: string;
    notification: string;
    timestamp: number;
}

export interface IPickData {
    data: {
        command: 'pick' | 'close'
        items: SPItem[];
        keepSharing: false;
    }
}

export interface PickerChangeEvent extends Event {
    detail: IPickData;
}

export interface PickerNotificationEvent extends Event {
    detail: INotificationData;
}

export interface PickerWindowEventMap {
    'pickerchange': PickerChangeEvent;
    'pickernotification': PickerNotificationEvent;
}

export type PickerWindow = {
    addEventListener<K extends keyof PickerWindowEventMap>(type: K, listener: (this: Window, ev: PickerWindowEventMap[K]) => any, options?: boolean | AddEventListenerOptions): void;
    removeEventListener<K extends keyof PickerWindowEventMap>(type: K, listener: (this: Window, ev: PickerWindowEventMap[K]) => any, options?: boolean | EventListenerOptions): void;
}

export interface BasePickerInit {
    options: IFilePickerOptions,
}

export interface OneDriveConsumerInit extends BasePickerInit {
    type: 'Consumer',
}

export interface ODSPInit extends BasePickerInit {
    type: 'ODSP',
    baseUrl: string,
}

export interface SPItem {
    '@sharePoint.embedUrl': string;
    '@sharePoint.endpoint': string;
    '@sharePoint.listUrl': string;
    folder?: any;
    id: string;
    name: string;
    parentReference: {
        driveId: string;
        sharepointIds: {
            listId: string;
            siteId: string;
            siteUrl: string;
            webId: string;
        }
    }
    sharepointIds: {
        listId: string;
        listItemId: string;
        listItemUniqueId: string;
        siteId: string;
        siteUrl: string;
        webId: string;
    }
    size: number;
    webDavUrl: string;
    webUrl: string;
}